requirements.txt
Flask
pandas
scikit-learn
matplotlib
joblib
pdfplumber
python-mysql-connector
werkzeug
